% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{combine_peaklists}
\alias{combine_peaklists}
\title{Read and Combine Bruker MSI peaklist files.}
\usage{
combine_peaklists(i.path, i.name = "peaklists", o.path = ".", o.name = NA)
}
\arguments{
\item{i.path}{Path to dataset of interest. This will usually be a folder with
a subfolder (\code{i.name}) that contains the peaklist files themselves.}

\item{i.name}{Name of the subfolder in \code{i.path} that contains the
peaklist files.}

\item{o.path}{Path to where output files should be written.}

\item{o.name}{Name identifying the dataset of interest, will default to the
data folder name provided in \code{i.path}.}
}
\value{
On successful completion teturns the number of empty spectra found
  --- peaklist files with a header but no peaks. If no peaklist files are
  found at all returns -1 and a warning.
}
\description{
Bruker mass spectrometry imaging software produces a folder filled with many
peaklists --- one for each spectrum. \code{combine_peaklists} reads all such
peaklist files in a given folder, and writes relevant information to two
tables: a peak-list, and a spectrum-list: \itemize{ \item Peak-list: A table
in which each row corresponds to a peak. Consists of all original peaklists
concatenated together with one additional column, \code{Acq}, identifying the
spectrum from which a peak originated. \item Spectrum-list: A table in which
each row corresponds to a spectrum. Columns contain information relevant at a
spectrum level, specifically this table contains five columns total:
\itemize{ \item \code{fname}: The original peaklist filename. \item \code{R}:
The region numbers of the corresponding spectra. \item \code{X}:     The
X-coordinates of the corresponding spectra. \item \code{Y}:     The
Y-coordinates of the corresponding spectra. \item \code{Acq}:   An integer
identifier used to cross-reference to the peak-list table. } } Note that
\code{combine_peaklists} checks for duplicate spectra (peaklist files with
the same region number, X-coordinate, and Y-coordinate) and will throw an
error if it finds any. Also note that \code{Acq} will correspond to the order
of acquisition of the spectra if the spectra where acquired in increasing
order of first region number, second Y-coordinate and third X-coordinate, in
that order of priority.
}
\examples{
combine_peaklists("A1")
}
\seealso{
\code{\link{load_peaklist}} and \code{\link{load_speclist}} for
  functions that can then load these resulting files once they have been
  created.
}

