% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dipps.R
\name{dipps_cutoff}
\alias{dipps_cutoff}
\title{Heuristic DIPPS cutoff}
\usage{
dipps_cutoff(df.peak, prop.m, id.var = "Acq", group.var = "group",
  subset.var = "subset")
}
\arguments{
\item{df.peak}{Parameter}

\item{prop.m}{The output from a call to \code{dipps}.}

\item{id.var}{Integer variable in \code{df.peak} that identifies unique
spectra.}

\item{group.var}{Integer variable in \code{df.peak} that identifies peaks
assigned to the same `peakgroup' -- typically grouped by m.z clustering.}

\item{subset.var}{Boolean variable in \code{df.peak} in which values of
TRUE identify the subset of spectra expected to be `upregulated'.}
}
\description{
Calculates a heuristic for the optimal DIPPS cutoff.
}
\seealso{
Winderbaum, L. J. et al. Feature extraction for proteomics imaging
  mass spectrometry data. The Annals of Applied Statistics.
  2015;9(4):1973-1996. doi: 10.1214/15-AOAS870.")
}

