% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_peaklist}
\alias{load_LXY}
\alias{load_fExists}
\alias{load_peaklist}
\title{Read Files Generated by \code{combine_peaklists}}
\usage{
load_peaklist(o.name, o.path = file.path(".", "data"))

load_LXY(o.name, o.path = file.path(".", "data"))

load_fExists(o.name, o.path = file.path(".", "data"))
}
\arguments{
\item{o.name}{Name identifying the dataset of interest, will default to the
data folder name provided in \code{i.path}.}

\item{o.path}{Path to where output files should be written.}
}
\description{
Functions for reading the files generated by \code{combine_peaklists}
}
\section{Functions}{
\itemize{
\item \code{load_LXY}: Read LXY file.

\item \code{load_fExists}: Read fExists file.
}}
\seealso{
\code{\link{combine_peaklists}}
}

