% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{tol_clus}
\alias{tol_clus}
\title{Tolerance Clustering}
\usage{
tol_clus(df.peak, tol = 0.1, var = "m.z")
}
\arguments{
\item{df.peak}{Parameter}

\item{tol}{Parameter}

\item{var}{Parameter}
}
\description{
Groups peaks together according to the equivalence classes induced by the
relation `are within tol of each other in mass'.
}
\examples{
\dontrun{
dataset.name = "A1" # Peaklist files should be in ./A1/peaklists/
n.empty = combine_peaklists(dataset.name)
df.peak = load_peaklist(dataset.name)
df.tol  = tol_clus(df.peak)
# Count how many peaks in each group
counts = table(df.tol$group)
# Remove peaks belonging to groups with less than 100 peaks total.
df.sub = subset(df.tol, group \%in\% as.numeric(names(counts)[counts >= 100]))
}
}

