% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{combine_peaklists}
\alias{combine_peaklists}
\title{Read and Combine Bruker MSI peaklist files.}
\usage{
combine_peaklists(i.path, i.name = "peaklists", o.path = file.path(".",
  "data"), o.name = NA)
}
\arguments{
\item{i.path}{Path to dataset of interest. This will usually be a folder with
a subfolder (\code{i.name}) that contains the peaklist files themselves.}

\item{i.name}{Name of the subfolder in \code{i.path} that contains the
peaklist files.}

\item{o.path}{Path to where output files should be written.}

\item{o.name}{Name identifying the dataset of interest, will default to the
data folder name provided in \code{i.path}.}
}
\value{
On successful completion teturns the number of empty spectra found
  --- peaklist files with a header but no peaks. If no peaklist files are
  found at all returns -1 and a warning.
}
\description{
Bruker mass spectrometry imaging software produces a folder filled with many
peaklists --- one for each spectrum. \code{combine_peaklists} reads all such
peaklist files in a given folder, and writes relevant information to three
tables that are useful and easier to handle downstream.
}
\details{
\code{combine_peaklists} writes three tables, \itemize{ \item
\code{Peaklist}: A peaklist consisting of all peaklists concatenated together
with one additional column, \code{Acq}, identifying the spectrum from which a
peak originated. \item \code{LXY}: A list of X-Y coordinates, or "LXY". In
this table rows represent spectra, i.e. seperate peaklist files in the Bruker
format. This table contains five columns: \itemize{ \item \code{fname}: The
original peaklist filename. \item \code{R}:     The region numbers of the
corresponding spectra. \item \code{X}:     The X-coordinates of the
corresponding spectra. \item \code{Y}:     The Y-coordinates of the
corresponding spectra. \item \code{Acq}:   An integer identifier used to
cross-reference to the peaklist table. } \item \code{fExists}: A matrix
representing a spatial grid identifying spatial X-Y coordinates where spectra
were acquired, i.e. for which peaklist files exist, or "fExists". If
\code{x.min} and \code{x.max} are the minimum and maximum X-coordinates at
which spectra were acquired, then and similarly \code{y.min} and \code{y.max}
are the minimum and maximum Y-coordinates, then the \code{fExists} matrix
represents a spatial grid spanning from \code{x.min - 1} to \code{x.max + 1}
on the X-axis, and from \code{y.min - 1} to \code{y.max + 1} on the Y-axis.
\code{fExists} contains zero values at spatial locations at which no spectra
was acquired (no peaklist file exists), and contains the corresponding value
of \code{Acq} at spatial locations at which spectra were acquired (peaklist
file exists). \code{fExists} is useful for constructing spatial plots TODO:
Add a See: here to plotting function. } Note that \code{combine_peaklists}
checked for duplicate spectra (peaklist files with the same region number,
X-coordinate, and Y-coordinate) and will throw an error if it finds any. Also
note that \code{Acq} will correspond to the order of acquisition of the
spectra if the spectra where acquired in increasing order of first region
number, second Y-coordinate and third X-coordinate, in that order of
priority.
}
\examples{
combine_peaklists("A1")
}
\seealso{
\code{\link{load_peaklist}}, \code{\link{load_LXY}}, and
  \code{\link{load_fExists}} for functions that can then load these resulting
  files once they have been created.
}

