% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{dbscan}
\alias{dbscan}
\title{Density Based Clustering}
\usage{
dbscan(df.peak, eps = 0.05, mnpts = 100, var = "m.z", pp = TRUE)
}
\arguments{
\item{df.peak}{Parameter}

\item{eps}{Parameter}

\item{mnpts}{Parameter}

\item{var}{Parameter}

\item{pp}{Parameter}
}
\description{
Groups peaks together according to the equivalence classes induced by the
relation `are within tol of each other in mass'.
}
\examples{
\dontrun{
dataset.name = "A1" # Peaklist files should be in ./A1/peaklists/
n.empty = combine_peaklists(dataset.name)
df.peak = load_peaklist(dataset.name)
df.tol  = dbscan(df.peak)
}
}
\seealso{
This is a univariate optimisation of the DBSCAN* algorithm as in
section 3 of:

Campello, Ricardo JGB, Davoud Moulavi, and Joerg Sander.
"Density-based clustering based on hierarchical density estimates." In
Advances in Knowledge Discovery and Data Mining, pp. 160-172. Springer Berlin
Heidelberg, 2013.
}

