% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{extract_masses}
\alias{extract_masses}
\title{Subsets a peaklist around known masses.}
\usage{
extract_masses(df.peak, mzs, margin = 0.3, use_ppm = FALSE, var = "m.z")
}
\arguments{
\item{df.peak}{Parameter}

\item{mzs}{Parameter}

\item{margin}{Parameter}

\item{use_ppm}{Parameter}

\item{var}{Parameter}
}
\description{
Subsets a peaklist around given masses with the given tolerance.
}

