% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_peaklist}
\alias{load_peaklist}
\alias{load_speclist}
\title{Read Files Generated by \code{combine_peaklists}}
\usage{
load_peaklist(o.name, o.path = ".")

load_speclist(o.name, o.path = ".")
}
\arguments{
\item{o.name}{Name identifying the dataset of interest. If left as \code{NA},
this will default to name of the data folder provided in \code{i.path}.}

\item{o.path}{Path to where output files should be written.}
}
\description{
Functions for reading the files generated by \code{combine_peaklists}
}
\section{Functions}{
\itemize{
\item \code{load_speclist}: Read LXY file.
}}
\seealso{
\code{\link{combine_peaklists}}
}

