% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{spatial_plot}
\alias{spatial_plot}
\title{Generate a \code{data.frame} for plotting with \code{ggplot2}}
\usage{
spatial_plot(df.peak, df.spec, plot.var = "intensity", id.var = "Acq",
  x.var = "X", y.var = "Y", return.plot = FALSE, print.plot = TRUE)
}
\arguments{
\item{df.peak}{Parameter}

\item{df.spec}{Parameter}

\item{plot.var}{Parameter}

\item{id.var}{Parameter}

\item{x.var}{Parameter}

\item{y.var}{Parameter}

\item{return.plot}{Parameter}

\item{print.plot}{Parameter}
}
\description{
Generate a \code{data.frame} for plotting with \code{ggplot2}
}
\examples{
\dontrun{
dataset.name = "A1"
combine_peaklists(dataset.name)
df.spec = load_speclist(dataset.name)
df.peak = load_peaklist(dataset.name)
# Select a m/z window of interest.
df.cal = subset(df.peak, abs(m.z - 1570.677) < 0.1)
if (sum(table(df.cal$Acq) > 1) > 0){
  stop(paste("multiple peaks from the same spectrum in given window.",
             "This should be dealt with before proceeding."))
}
df.cal$log.intensity = log1p(df.cal$intensity)
# Plot log-intensity of said m/z window.
df.plot = spatial_plot(df.cal, df.spec, plot.var = "log.intensity")
}
}

